requirements.txt
streamlit
ethos-u-vela
supabase==2.27.0
python-dotenv==1.0.1
websockets==15.0.1
gotrue
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24